pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crispr-editing-env"
version = "2.0.0"
description = "CRISPR gene editing environment with bioinformatics tool-use for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.0,<3.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "openai>=1.12.0",
    "openenv-core[core]>=0.2.2",
]

[project.scripts]
server = "crispr_editing_env.server.app:main"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]

[tool.setuptools]
packages = ["crispr_editing_env", "crispr_editing_env.server"]

[tool.setuptools.package-dir]
"crispr_editing_env" = "."
"crispr_editing_env.server" = "server"
